refactor(user): extract user module into dojo/user/ [user Phase 1,3,4,5,6,7,8,9]
