+++ dojo/user/__init__.py
@@ -1,0 +1,1 @@
+import dojo.user.admin  # noqa: F401
