+++ dojo/user/api/__init__.py
@@ -1,0 +1,1 @@
+path = "users"  # noqa: RUF067
